♻️ refactor:  stateフォルダを廃止してcomponentsフォルダに統合
Issue #157
