+++ build.yaml
@@ -9,15 +9,15 @@ targets:
         generate_for:
           include:
             - lib/domain/**/*.dart
-            - lib/**/json_objects/*.dart
-            - lib/**/state/*.dart
+            - lib/infrastructure/**/json_objects/*.dart
+            - lib/presentation/**/*.dart
             - lib/**/*state.dart
       json_serializable:
         generate_for:
           include:
             - lib/domain/**/*.dart
-            - lib/**/json_objects/*.dart
-            - lib/**/state/*.dart
+            - lib/infrastructure/**/json_objects/*.dart
+            - lib/presentation/**/*.dart
             - lib/**/*state.dart
       fast_i18n:
         options:
