+++ lib/presentation/common/components/launch_url_state.freezed.dart
