+++ lib/presentation/repo/components/avatar_preview_view.dart
@@ -8,7 +8,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:photo_view/photo_view.dart';
 
 import '../../../domain/entities/repo.dart';
-import '../state/selected_repo.dart';
+import 'selected_repo.dart';
 
 /// アバター画像プレビューView
 class RepoAvatarPreviewView extends ConsumerWidget {
