+++ lib/presentation/repo/components/sort_selector_bottom_sheet.dart
@@ -8,8 +8,8 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../../../domain/entities/search_repos_sort.dart';
 import '../../../localizations/strings.g.dart';
 import '../../../utils/logger.dart';
-import '../state/search_repos_sort.dart';
 import 'order_toggle_button.dart';
+import 'search_repos_sort.dart';
 
 /// リポジトリ検索用ソート選択ボトムシート
 class RepoSortSelectorBottomSheet extends ConsumerWidget {
