+++ test/presentation/common/components/launch_url_state_test.dart
@@ -4,7 +4,7 @@
 
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:github_search/presentation/common/state/launch_url_state.dart';
+import 'package:github_search/presentation/common/components/launch_url_state.dart';
 import 'package:mocktail/mocktail.dart';
 // ignore: depend_on_referenced_packages
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
