+++ test/presentation/repo/components/order_toggle_button_test.dart
@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:github_search/infrastructure/github/http_client.dart';
 import 'package:github_search/presentation/repo/components/order_toggle_button.dart';
-import 'package:github_search/presentation/repo/state/search_repos_query.dart';
+import 'package:github_search/presentation/repo/components/search_repos_query.dart';
 
 import '../../../test_utils/logger.dart';
