+++ test/presentation/repo/components/search_text_button_test.dart
@@ -4,10 +4,10 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:github_search/presentation/repo/components/search_repos_query.dart';
 import 'package:github_search/presentation/repo/components/search_text_button.dart';
 import 'package:github_search/presentation/repo/components/search_text_field.dart';
 import 'package:github_search/presentation/repo/pages/search_page.dart';
-import 'package:github_search/presentation/repo/state/search_repos_query.dart';
 import 'package:go_router/go_router.dart';
 import 'package:mocktail/mocktail.dart';
