+++ test/test_utils/test_agent.dart
@@ -20,7 +20,7 @@ import 'package:github_search/infrastructure/hive/app_data_repository.dart';
 import 'package:github_search/infrastructure/isar/collections/query_history.dart';
 import 'package:github_search/infrastructure/isar/query_history_repository.dart';
 import 'package:github_search/localizations/strings.g.dart';
-import 'package:github_search/presentation/repo/state/search_repos_query.dart';
+import 'package:github_search/presentation/repo/components/search_repos_query.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 import 'package:isar/isar.dart';
 // ignore: unused_import
